requirements.txt
accelerate==1.4.0
acres==0.3.0
aiofiles==24.1.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
Django==5.1.7
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.17.0
fitz==0.0.1.dev2
frontend==0.0.3
fsspec==2025.3.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.29.3
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
looseversion==1.3.0
lxml==5.3.1
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.1.0
prov==2.0.1
psutil==7.0.0
puremagic==1.28
pydot==3.0.4
PyMuPDF==1.25.4
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytils==0.4.1
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
sympy==1.13.1
tiktoken==0.9.0
tokenizers==0.21.0
tools==0.1.9
torch==2.6.0
tqdm==4.67.1
traits==7.0.2
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
webvtt-py==0.5.1
whisper==1.1.10
yt-dlp==2025.2.19
